Move generic comms methods from `TorchTensor` to `AbstractTensor`
Some of the methods on `TorchTensor` don't have any references to Torch
or to other concrete tensor classes, so they can be directly moved up
the inheritance hierarchy to `AbstractTensor`. (Those that don't
reference tensors may be able to move farther up to `AbstractSendable`.)

This allows some of the methods to be marked as `@abstractmethods` to
require them to be implemented by all child classes (since they are now.)
